New Level
Made Board 3 (Went a bit overboard with the bad holes, lmk iif I should tone it down a bit) and made the prefab and tag for the bad holes. They don't have a script yet, but you can just tack that onto the prefab and it should apply to all the bad holes already on the board.